Replaced MD5 with SHA256 #2116
